+++ package.json
@@ -2,10 +2,11 @@
   "name": "mcp-github-oauth",
   "version": "0.0.1",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "tsc",
-    "deploy": "wrangler deploy",
-    "dev": "wrangler dev",
+    "build": "tsc -b && vite build",
+    "dev": "vite",
+    "deploy": "npm exec wrangler deploy",
     "eval": "evalite",
     "eval:ci": "evalite --threshold=100",
     "eval:dev": "evalite watch",
@@ -13,7 +14,7 @@
     "lint": "biome lint",
     "lint:fix": "biome lint --fix",
     "inspector": "pnpx @modelcontextprotocol/inspector@latest",
-    "start": "wrangler dev",
+    "preview": "vite preview",
     "start:stdio": "tsx src/mcp/start-stdio.ts",
     "cf-typegen": "wrangler types",
     "postinstall": "simple-git-hooks",
@@ -25,16 +26,25 @@
   "devDependencies": {
     "@ai-sdk/openai": "^1.3.7",
     "@biomejs/biome": "^1.9.4",
+    "@cloudflare/vite-plugin": "^1.0.1",
     "@cloudflare/vitest-pool-workers": "^0.8.11",
+    "@cloudflare/workers-types": "^4.20250407.0",
+    "@tailwindcss/typography": "^0.5.16",
+    "@tailwindcss/vite": "^4.1.3",
     "@types/node": "^22.14.0",
+    "@types/react": "^19.1.0",
+    "@types/react-dom": "^19.1.1",
+    "@vitejs/plugin-react": "^4.3.4",
     "@vitest/coverage-v8": "^2.1.9",
     "ai": "^4.3.2",
     "autoevals": "^0.0.125",
+    "better-sqlite3": "^11.9.1",
     "dotenv": "^16.4.7",
     "evalite": "~0.10.0",
     "lint-staged": "^15.5.0",
     "msw": "^2.7.3",
     "simple-git-hooks": "^2.12.1",
+    "tailwindcss": "^4.1.3",
     "tsx": "^4.19.3",
     "typescript": "^5.8.3",
     "vite": "^6.2.5",
@@ -46,8 +56,9 @@
     "@modelcontextprotocol/sdk": "^1.8.0",
     "@sentry/cloudflare": "^9.11.0",
     "agents": "~0.0.54",
-    "better-sqlite3": "^11.9.1",
     "hono": "^4.7.5",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "workers-mcp": "0.1.0-3",
     "zod": "^3.24.2"
   },
