+++ src/mcp/tools.ts
@@ -1,17 +1,10 @@
-import type { ServerContext } from "../types";
-import { z } from "zod";
+import type { z } from "zod";
 import type {
   SentryErrorEntrySchema,
   SentryEventSchema,
 } from "../lib/sentry-api";
 import { SentryApiService, extractIssueId } from "../lib/sentry-api";
-import {
-  ParamIssueShortId,
-  ParamOrganizationSlug,
-  ParamPlatform,
-  ParamTeamSlug,
-} from "./schema";
-import type { RequestHandlerExtra } from "@modelcontextprotocol/sdk/shared/protocol.js";
+import type { ToolHandlers } from "./types";
 
 function formatEventOutput(event: z.infer<typeof SentryEventSchema>) {
   let output = "";
@@ -46,146 +39,6 @@ function formatEventOutput(event: z.infer<typeof SentryEventSchema>) {
   }
   return output;
 }
-
-export const TOOL_DEFINITIONS = [
-  {
-    name: "list_organizations" as const,
-    description: [
-      "List all organizations that the user has access to in Sentry.",
-      "",
-      "Use this tool when you need to:",
-      "- View all organizations in Sentry",
-    ].join("\n"),
-  },
-  {
-    name: "list_teams" as const,
-    description: [
-      "List all teams in an organization in Sentry.",
-      "",
-      "Use this tool when you need to:",
-      "- View all teams in a Sentry organization",
-    ].join("\n"),
-    paramsSchema: {
-      organizationSlug: ParamOrganizationSlug.optional(),
-    },
-  },
-  {
-    name: "list_projects" as const,
-    description: [
-      "Retrieve a list of projects in Sentry.",
-      "",
-      "Use this tool when you need to:",
-      "- View all projects in a Sentry organization",
-    ].join("\n"),
-    paramsSchema: {
-      organizationSlug: ParamOrganizationSlug.optional(),
-    },
-  },
-  {
-    name: "get_error_details" as const,
-    description: [
-      "Retrieve error details from Sentry for a specific Issue ID, including the stacktrace and error message. Either issueId or issueUrl MUST be provided.",
-      "",
-      "Use this tool when you need to:",
-      "- Investigate a specific production error",
-      "- Access detailed error information and stacktraces from Sentry",
-    ].join("\n"),
-    paramsSchema: {
-      organizationSlug: ParamOrganizationSlug.optional(),
-      issueId: ParamIssueShortId.optional(),
-      issueUrl: z
-        .string()
-        .url()
-        .describe("The URL of the issue to retrieve details for.")
-        .optional(),
-    },
-  },
-  {
-    name: "search_errors_in_file" as const,
-    description: [
-      "Search for errors recently occurring in a specific file. This is a suffix based search, so only using the filename or the direct parent folder of the file. The parent folder is preferred when the filename is in a subfolder or a common filename.",
-      "",
-      "Use this tool when you need to:",
-      "- Search for production errors in a specific file",
-      "- Analyze error patterns and frequencies",
-      "- Find recent or frequently occurring errors.",
-    ].join("\n"),
-    paramsSchema: {
-      organizationSlug: ParamOrganizationSlug.optional(),
-      filename: z.string().describe("The filename to search for errors in."),
-      sortBy: z
-        .enum(["last_seen", "count"])
-        .optional()
-        .default("last_seen")
-        .describe(
-          "Sort the results either by the last time they occurred or the count of occurrences.",
-        ),
-    },
-  },
-  {
-    name: "create_team" as const,
-    description: [
-      "Create a new team in Sentry.",
-      "",
-      "Use this tool when you need to:",
-      "- Create a new team in a Sentry organization",
-    ].join("\n"),
-    paramsSchema: {
-      organizationSlug: ParamOrganizationSlug.optional(),
-      name: z.string().describe("The name of the team to create."),
-    },
-  },
-  {
-    name: "create_project" as const,
-    description: [
-      "Create a new project in Sentry, giving you access to a new SENTRY_DSN.",
-      "",
-      "Use this tool when you need to:",
-      "- Create a new project in a Sentry organization",
-    ].join("\n"),
-    paramsSchema: {
-      organizationSlug: ParamOrganizationSlug.optional(),
-      teamSlug: ParamTeamSlug,
-      name: z
-        .string()
-        .describe(
-          "The name of the project to create. Typically this is commonly the name of the repository or service. It is only used as a visual label in Sentry.",
-        ),
-      platform: ParamPlatform.optional(),
-    },
-  },
-];
-
-export type ToolName = (typeof TOOL_DEFINITIONS)[number]["name"];
-
-export type ToolDefinition<T extends ToolName> = Extract<
-  (typeof TOOL_DEFINITIONS)[number],
-  { name: T }
->;
-
-type ZodifyRecord<T extends Record<string, any>> = {
-  [K in keyof T]: z.infer<T[K]>;
-};
-
-export type ToolParams<T extends ToolName> = ToolDefinition<T> extends {
-  paramsSchema: Record<string, any>;
-}
-  ? ZodifyRecord<ToolDefinition<T>["paramsSchema"]>
-  : Record<string, never>;
-
-export type ToolHandler<T extends ToolName> = (
-  params: ToolParams<T>,
-) => Promise<string>;
-
-export type ToolHandlerExtended<T extends ToolName> = (
-  context: ServerContext,
-  params: ToolParams<T>,
-  extra: RequestHandlerExtra,
-) => Promise<string>;
-
-export type ToolHandlers = {
-  [K in ToolName]: ToolHandlerExtended<K>;
-};
 
 export const TOOL_HANDLERS = {
   list_organizations: async (context) => {
