+++ src/mcp/transports/cloudflare-worker.ts
@@ -1,7 +1,7 @@
 import { McpAgent } from "agents/mcp";
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import type { WorkerProps } from "../../types";
 import { configureServer } from "../server";
+import type { Env, WorkerProps } from "../../worker/types";
 
 // Context from the auth process, encrypted & stored in the auth token
 // and provided to the DurableMCP as this.props
