--- src/types.ts
@@ -1,9 +1,0 @@
-export type ServerContext = {
-  accessToken: string;
-  organizationSlug: string | null;
-};
-
-export type WorkerProps = ServerContext & {
-  id: string;
-  name: string;
-};
