+++ tsconfig.json
@@ -1,29 +1,8 @@
 {
-  "include": ["src/**/*", "worker-configuration.d.ts", "test/**/*"],
-  "exclude": ["node_modules"],
-  "compilerOptions": {
-    "target": "es2021",
-    "lib": ["es2021"],
-    "jsx": "react-jsx",
-    "jsxImportSource": "hono/jsx",
-
-    "rootDir": "./src",
-    "outDir": "./dist",
-
-    "module": "es2022",
-    "moduleResolution": "Bundler",
-    "types": ["@cloudflare/vitest-pool-workers"],
-    "resolveJsonModule": true,
-
-    "allowJs": true,
-    "checkJs": false,
-
-    "isolatedModules": true,
-    "allowSyntheticDefaultImports": true,
-    "forceConsistentCasingInFileNames": true,
-
-    "strict": true,
-
-    "skipLibCheck": true
-  }
+  "files": [],
+  "references": [
+    { "path": "./tsconfig.app.json" },
+    { "path": "./tsconfig.node.json" },
+    { "path": "./tsconfig.worker.json" }
+  ]
 }
