+++ worker-configuration.d.ts
@@ -1,13 +1,13 @@
-// Generated by Wrangler by running `wrangler types` (hash: d649a02a693e4378c75cd07db6b92680)
+// Generated by Wrangler by running `wrangler types` (hash: 98fb12cef628dd54385b9692782d2798)
 // Runtime types generated with workerd@1.20250404.0 2025-03-21 nodejs_compat
 declare namespace Cloudflare {
 	interface Env {
 		OAUTH_KV: KVNamespace;
+		SENTRY_DSN: "https://7f7bbaad9504b727cdf8edc378c6d1de@o1.ingest.us.sentry.io/4509062593708032";
 		SENTRY_URL: "https://sentry.io";
 		SENTRY_CLIENT_ID: string;
 		SENTRY_CLIENT_SECRET: string;
-		SENTRY_DSN: string;
-		MCP_OBJECT: DurableObjectNamespace<import("./src/index").SentryMCP>;
+		MCP_OBJECT: DurableObjectNamespace<import("./src/worker/index").SentryMCP>;
 		AI: Ai;
 	}
 }
